refactor(rooms): split item persistence
